requirements.txt
Pyrebase4==4.4.1
flask
waitress